% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_lik_ICL_export.R
\name{comp_lik_ICL_export}
\alias{comp_lik_ICL_export}
\title{compute the Integrated likeilhood and the ICL criteria for the MBM}
\usage{
comp_lik_ICL_export(tau, pi, alpha, dataR6)
}
\arguments{
\item{tau}{Posterior probabilities of clustering}

\item{pi}{Type of the matrix, choice between "inc", "adj" and "diradj"}

\item{alpha}{Name of the functional group in row}

\item{colFG}{Name of the function group in column}
}
\value{
a list object formatted for the GREMLIN package
}
\description{
compute the Integrated likeilhood and the ICL criteria for the MBM
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
DefineNetwork(A,"diradj","FG1","FG1")
}
