% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparClassif.R
\name{comparClassif}
\alias{comparClassif}
\title{Compare two classifications on all the  Functional groups}
\usage{
comparClassif(classif1, classif2)
}
\arguments{
\item{classif1}{: list a length n_FG.}

\item{classif2}{: list a length n_FG.}
}
\value{
Adjusted Rand Index (ARI) for each Functional Group.
}
\description{
Compare two classifications on all the  Functional groups
}
\examples{
nFG <- 3;
vK <- c(4,5,2) ;
vNQ <- c(100,40,50);
classif1 <- lapply(1:nFG,function(q){sample(1:vK[q],vNQ[q],replace=TRUE)})
classif2 <- classif1
classif2[[2]] <-  sample(1:vK[2],vNQ[2],replace=TRUE)
resCompar <- comparClassif (classif1,classif2)
}
