% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compLikICL.R
\name{compLikICL}
\alias{compLikICL}
\title{compute the Integrated likeilhood and the ICL criteria for the MBM}
\usage{
compLikICL(paramEstim, list_Net, v_distrib = NULL)
}
\arguments{
\item{paramEstim}{Estimated parameters of MBM}

\item{list_Net}{A list of network}

\item{v_distrib}{Type of proababilistic distributions in each network : if 0/1 then Bernoulli, if counting then Poisson. My default  = Bernoulli.
Must give a vector whose length is the number of networks in list_Net}
}
\value{
Pseudo-Likelihood and penalty
}
\description{
compute the Integrated likeilhood and the ICL criteria for the MBM
}
