% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compLikICLExport.R
\name{compLikICLExport}
\alias{compLikICLExport}
\title{compute the Integrated likeilhood and the ICL criteria for the MBM}
\usage{
compLikICLExport(paramEstim, list_Net)
}
\arguments{
\item{list_Net}{A list of network}

\item{param_est}{Estimated parameters of MBM}
}
\value{
Pseudo-Likelihood, penalty
}
\description{
compute the Integrated likeilhood and the ICL criteria for the MBM
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
defineNetwork(A,"diradj","FG1","FG1")
}
