% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineNetwork.R
\name{DefineNetwork}
\alias{DefineNetwork}
\title{Define a network}
\usage{
DefineNetwork(mat, type, rowFG, colFG)
}
\arguments{
\item{mat}{An adjacency (symmetric or not) or an incidence matrix}

\item{type}{Type of the matrix, choice between "inc", "adj" and "diradj"}

\item{rowFG}{Name of the functional group in row}

\item{colFG}{Name of the function group in column}
}
\value{
a list object formatted for the GREMLIN package
}
\description{
Define a network
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
DefineNetwork(A,"diradj","FG1","FG1")
}
