% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMBM.R
\name{plotMBM}
\alias{plotMBM}
\title{Plot the mesoscopic view  of the estimated MBM}
\usage{
plotMBM(resMBM, whichModel = 1, mycol = NULL, thres = 0.01, maxCurved = 3)
}
\arguments{
\item{resMBM}{A fitted Generalized BlockModel}

\item{whichModel}{The index corresponding to the model to plot (default is 1, the best model)}

\item{mycol}{A list of colors for the functional groups}

\item{thres}{A threshold under which edges correponding to probability of connections are not plotted}

\item{maxCurved}{graphical parameter : curvature of the edges}
}
\description{
Plot the mesocopic view of the multipartite network obtained by the Genreliazed block models. resMBM is the results of of MBM fitting (output of multipartiteBMFixedModel for given numbers of clusters or multipartiteBM if the number of clusters is selected)
}
\examples{
namesFG <- c('A','B')
list_pi = list(c(0.16 ,0.40 ,0.44),c(0.3,0.7))
E  <-  rbind(c(1,2),c(2,2),c(1,1))
typeInter <- c( "inc","diradj", "adj")
v_distrib <- c('gaussian','bernoulli','poisson')
list_theta <- list()
list_theta[[1]] <- list()
list_theta[[1]]$mean  <- matrix(c(6.1, 8.9, 6.6, 9.8, 2.6, 1.0), 3, 2)
list_theta[[1]]$var  <-  matrix(c(1.6, 1.6, 1.8, 1.7 ,2.3, 1.5),3, 2)
list_theta[[2]] <- matrix(c(0.7,1.0, 0.4, 0.6),2, 2)
m3 <- matrix(c(2.5, 2.6 ,2.2 ,2.2, 2.7 ,3.0 ,3.6, 3.5, 3.3),3,3 )
list_theta[[3]] <- (m3 + t(m3))/2
list_Net <- rMBM(v_NQ = c(60,50),E , typeInter, v_distrib, list_pi,
                list_theta, namesFG = namesFG, seed = 2)$list_Net
res_MBMsimu <- multipartiteBM(list_Net, v_distrib,
                             namesFG = c('A','B'), v_Kinit = c(2,2),
                             nbCores = 2,initBM = FALSE)
plotMBM(res_MBMsimu)
}
