% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_lik_ICL_export.R
\name{comp_lik_ICL_export}
\alias{comp_lik_ICL_export}
\title{compute the Integrated likeilhood and the ICL criteria for the MBM}
\usage{
comp_lik_ICL_export(param_estim, listNet)
}
\arguments{
\item{listNet}{A list of network}

\item{param_est}{Estimated parameters of MBM}
}
\value{
Pseudo-Likelihood, penalty
}
\description{
compute the Integrated likeilhood and the ICL criteria for the MBM
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
DefineNetwork(A,"diradj","FG1","FG1")
}
