% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compLikICL.R
\name{compLikICL}
\alias{compLikICL}
\title{compute the Integrated likeilhood and the ICL criteria for the MBM}
\usage{
compLikICL(paramEstim, list_Net, v_distrib = NULL)
}
\arguments{
\item{list_Net}{A list of network}

\item{param_est}{Estimated parameters of MBM}
}
\value{
Pseudo-Likelihood, penalty
}
\description{
compute the Integrated likeilhood and the ICL criteria for the MBM
}
\examples{

}
