% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineNetwork.R
\name{defineNetwork}
\alias{defineNetwork}
\title{Define a network providing its matrix of interactions and specifying the functions groups in row and col.}
\usage{
defineNetwork(mat, typeInter, rowFG, colFG)
}
\arguments{
\item{mat}{An adjacency matrix (symmetric or not) or an incidence matrix}

\item{typeInter}{Type of the matrix, choice between "inc" (incidence), "adj" (adjacency) and "diradj" (directed adjacency)}

\item{rowFG}{Name of the functional group in row}

\item{colFG}{Name of the function group in column}
}
\value{
a list object formatted for the GREMLIN package
}
\description{
Define a network providing its matrix of interactions and specifying the functions groups in row and col.
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
defineNetwork(A,"diradj","FG1","FG1")
}
