% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClustersMBM.R
\name{extractClustersMBM}
\alias{extractClustersMBM}
\title{Extract the clusters in each functional group}
\usage{
extractClustersMBM(resMBM, whichModel = 1)
}
\arguments{
\item{resMBM}{A fitted Generalized BlockModel}

\item{whichModel}{The index corresponding to the model to plot (default is 1, the best model)}
}
\value{
a list a length the number of Functional Groups. Each element is a list of length the number of blocks composed of the index of the individuals in each block of each cluster.
}
\description{
Extract the clusters in each functional group
}
