% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipartiteBM.R
\name{MultipartiteBM}
\alias{MultipartiteBM}
\title{Model selection and estimation of multipartite blockmodels}
\usage{
MultipartiteBM(listNet, namesFG = NULL, vKmin = 1, vKmax = 10,
  vKinit = NULL, verbose = TRUE, save = FALSE)
}
\arguments{
\item{listNet}{A list of network (defined via the function DefineNetwork)}

\item{namesFG}{Names of functional groups (must correspond to names in listNet)}

\item{vKmin}{A vector of minimal number of blocks per functional group provided in the same order as in namesFG}

\item{vKmax}{A vector of maximal number of blocks per functional group provided in the same order as in namesFG}

\item{vKinit}{A vector of initial number of blocks per functional group provided in the same order as in namesFG}

\item{verbose}{Set to TRUE to display the current step of the search algorithm}

\item{save}{Set to TRUE to save the estimated parameters for intermediate visited models}
}
\value{
a list of estimated parameters for the different models
}
\description{
Select the number of blocks per functional group using a stepwise search and estimate parameters
}
\examples{
npc1 <- 20 # nodes per class
Q1 <- 3 # classes
n1 <- npc1 * Q1 # nodes
Z1 <- diag(Q1)\%x\%matrix(1,npc1,1)
P1 <- matrix(runif(Q1*Q1),Q1,Q1)
A <- 1*(matrix(runif(n1*n1),n1,n1)<Z1\%*\%P1\%*\%t(Z1)) ## adjacency matrix
Agr <- DefineNetwork(A,"diradj","FG1","FG1")
npc2 <- 30 # nodes per class
Q2 <- 2 # classes
n2 <- npc2 * Q2 # nodes
Z2 <- diag(Q2)\%x\%matrix(1,npc2,1)
P2 <- matrix(runif(Q1*Q2),Q1,Q2)
B <- 1*(matrix(runif(n1*n2),n1,n2)<Z1\%*\%P2\%*\%t(Z2)) ## adjacency matrix
Bgr <- DefineNetwork(B,"inc","FG1","FG2")
res <- MultipartiteBM(list(Agr,Bgr),namesFG = NULL,vKmin = 1,vKmax = 10,vKinit = NULL,verbose = TRUE, save=FALSE)
}
