% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipartiteBMFixedModel.R
\name{multipartiteBMfixedmodel}
\alias{multipartiteBMfixedmodel}
\title{Model selection and estimation of multipartite blockmodels}
\usage{
multipartiteBMfixedmodel(list_Net, namesFG, v_K = NULL, classifInit = NULL,
  nbCores = NULL)
}
\arguments{
\item{list_Net}{A list of network (defined via the function DefineNetwork)}

\item{namesFG}{Names of functional groups (must correspond to names in listNet)}

\item{vK}{A vector with the numbers of blocks per functional group}

\item{classif.init}{A list of initial classification for each functional group in the same order as in namesFG}

\item{nb_cores}{Number of cores used for estimation}
}
\value{
Estimated parameters and a classification
}
\description{
Select the number of blocks per functional group using a stepwise search and estimate parameters
}
\examples{
npc1 <- 20 # nodes per class
Q1 <- 3 # classes
n1 <- npc1 * Q1 # nodes
Z1 <- diag(Q1)\%x\%matrix(1,npc1,1)
P1 <- matrix(runif(Q1*Q1),Q1,Q1)
A <- 1*(matrix(runif(n1*n1),n1,n1)<Z1\%*\%P1\%*\%t(Z1)) ## adjacency matrix
Agr <- DefineNetwork(A,"diradj","FG1","FG1")
npc2 <- 30 # nodes per class
Q2 <- 2 # classes
n2 <- npc2 * Q2 # nodes
Z2 <- diag(Q2)\%x\%matrix(1,npc2,1)
P2 <- matrix(runif(Q1*Q2),Q1,Q2)
B <- 1*(matrix(runif(n1*n2),n1,n2)<Z1\%*\%P2\%*\%t(Z2)) ## incidence matrix
Bgr <- defineNetwork(B,"inc","FG1","FG2")
res <- multipartiteBMFixedModel(list(Agr,Bgr),namesFG=c("FG1","FG2"),vK=c(3,2))
}
