% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMBM.R
\name{predictMBM}
\alias{predictMBM}
\title{Predict NAs in a Collection of Networks from a fitted MBM}
\usage{
predictMBM(RESMBM)
}
\arguments{
\item{RESMBM}{a fitted multipartite blockmodel}
}
\value{
the collection of networks where the NAs are replaced with predictions (only for Bernoulli model yet) and
when there are NAs an additional field \code{wNA} indicating where are the NAs.
}
\description{
Predict NAs in a Collection of Networks from a fitted MBM
}
\details{
works only for Bernoulli distribution on all networks. Be sure to define adjacency matrices with NA on the
diagonal if there are no loop.
}
