% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMBM.R
\name{rMBM}
\alias{rMBM}
\title{Simulates multipartite blocl models}
\usage{
rMBM(v_NQ, E, type_inter, vdistrib, lpi, ltheta, seed = NULL,
  namesfg = NULL)
}
\arguments{
\item{v_NQ}{: number of individual in each Functional Group (FG)}

\item{vdistrib}{: vector of the distributions  (bernoulli or poisson for each network)}

\item{lpi}{: parameters of the clustering distribution}

\item{ltheta}{: parameters of the interactions distribution}

\item{namesfg}{: names of the FG.}

\item{mat_E}{: define the architecture of the Multipartite.}

\item{type_int}{: type of interaction (  adjacency symetric or not or incidence)}
}
\value{
Pseudo-Likelihood, penalty
}
\description{
Simulates multipartite blocl models
}
\examples{
A <- matrix(rbinom(100,1,.2),10,10)
type <- "diradj"
DefineNetwork(A,"diradj","FG1","FG1")
}
