% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMBM.R
\name{rMBM}
\alias{rMBM}
\title{Simulate datasets from  multipartite block  models}
\usage{
rMBM(v_NQ, E, type_inter, vdistrib, lpi, ltheta, seed = NULL,
  namesfg = NULL)
}
\arguments{
\item{v_NQ}{: number of individual in each Functional Group (FG)}

\item{E}{: define the architecture of the Multipartite.}

\item{type_inter}{: type of interaction (  adjacency symetric or not or incidence) (vector of size equal to nrow(E) )}

\item{vdistrib}{: vector of the distributions  (bernoulli or poisson for each network) ( vector of size equal to nrow(E) )}

\item{lpi}{: parameters of the blocks distribution}

\item{ltheta}{: parameters of the interactions distribution}

\item{seed}{: set the seed for the random simulation (default value  = NULL)}

\item{namesfg}{: names of the FG.  (default value  = NULL, then the functional groups are labelled FG1, FG2, etc)}
}
\value{
A list of lists. Each element of the list corresponds to a network : each network is described by a matrix (mat) , a type (diradj, adj, inc), the functional group in row (rowFG) and the functional group in columns (colFG)
}
\description{
Simulate datasets from  multipartite block  models
}
\examples{
v_K <- c(3,2,2)
n_FG <- 3
lpi <- vector("list", 3);
lpi[[1]] <- c(0.4,0.3,0.3); lpi[[2]] <- c(0.6,0.4); lpi[[3]]  <- c(0.6,0.4)
E  = rbind(c(1,2),c(2,3),c(2,2))
vdistrib <- c('bernoulli','poisson','poisson')
type_inter <- c( "inc", "inc"  ,  "adj" )
ltheta <- list()
ltheta[[1]] <- matrix(rbeta(v_K[E[1,1]] * v_K[E[1,2]],1.5,1.5 ),nrow = v_K[E[1,1]], ncol = v_K[E[1,2]] )
ltheta[[2]] <- matrix(rgamma(v_K[E[2,1]] * v_K[E[2,2]],7.5,1 ),nrow = v_K[E[2,1]], ncol = v_K[E[2,2]] )
ltheta[[3]] <- matrix(rgamma(v_K[E[3,1]] * v_K[E[3,2]],7.5,1 ),nrow = v_K[E[3,1]], ncol = v_K[E[3,2]] )
ltheta[[3]] <- 0.5*(ltheta[[3]] + t(ltheta[[3]])) # symetrisation for network 3
v_NQ = c(100,50,40)
data_sim <- rMBM(v_NQ ,E , type_inter, vdistrib, lpi, ltheta, seed=NULL, namesfg= c('A','B','D'))
}
